Refactored treeview services (#7)
